refactor: Delete unused infomartion and dependencies
