+++ apps/src/main/kotlin/com/github/daitasu/Application.kt
@@ -1,22 +1,9 @@
 package com.github.daitasu
 
 import org.apache.pekko.actor.typed.ActorSystem
-import org.apache.pekko.actor.typed.Behavior
-import org.apache.pekko.actor.typed.javadsl.Behaviors
 import org.apache.pekko.cluster.typed.Cluster
 import org.apache.pekko.cluster.typed.Join
 import com.typesafe.config.ConfigFactory
-
-data class GreetingMessage(val message: String)
-
-object GreetingActor {
-  fun create(): Behavior<GreetingMessage> =
-    Behaviors.receive<GreetingMessage> { context, message ->
-      val member = Cluster.get(context.system).selfMember()
-      println("[Address: ${member.address()}] Message: ${message.message}")
-      Behaviors.same<GreetingMessage>()
-  }
-}
 
 fun main() {
     val port = System.getenv("CLUSTER_PORT")?.toIntOrNull() ?: 2551
