+++ apps/src/main/resources/application.conf
@@ -10,9 +10,11 @@ pekko {
   }
   remote {
     log-remote-lifecycle-events = on
-    artery.canonical {
-      hostname = ${clustering.ip}
-      port = ${clustering.port}
+    artery {
+      canonical {
+        hostname = ${clustering.ip}
+        port = ${clustering.port}
+      }
     }
   }
   cluster {
@@ -23,11 +25,11 @@ pekko {
 }
 
 clustering {
- ip = "127.0.0.1"
+ ip = seed
  ip = ${?CLUSTER_IP}
  port = 2551
  port = ${?CLUSTER_PORT}
- seed-ip = "127.0.0.1"
+ seed-ip = seed
  seed-port = 2551
- cluster.name = clustering-cluster
+ cluster.name = MyActorSystem
 }
