+++ docker/Dockerfile
@@ -1,11 +1,3 @@
-FROM gradle:8.12.0-jdk17 AS development
-
-WORKDIR /workspace
-
-EXPOSE 2551
-
-CMD ["gradle", "run"]
-
 FROM gradle:8.12.0-jdk17 AS builder
 
 WORKDIR /workspace
